refactor: move danish zone
