+++ libs/co2/data-access/src/lib/co2-emissions/store/create-co2-forecast-interval.spec.ts
@@ -1,6 +1,6 @@
 import { DateTime } from 'luxon';
 import { createCo2ForecastInterval } from './create-co2-forecast-interval';
-import { danishZone } from './danish-zone';
+import { danishZone } from '../date-time-util/danish-zone';
 
 describe(createCo2ForecastInterval.name, () => {
   it('returns the start of today as the date query from', () => {
