+++ libs/co2/data-access/src/lib/co2-emissions/store/create-co2-forecast-interval.ts
@@ -1,5 +1,5 @@
 import { DateTime, Duration, Interval } from 'luxon';
-import { danishZone } from './danish-zone';
+import { danishZone } from '../date-time-util/danish-zone';
 export function createCo2ForecastInterval(now: DateTime): Interval {
   const danishToday = now.setZone(danishZone).startOf('day');
   const danishDayAfterTomorrow = danishToday.plus(Duration.fromISO('P2D'));
